refactor(notification): mv delivery status
